fix: Download posters from CloudModel
Progress bar does not load completely until view is fully scrolled, because
it needs to wait for the RecyclerView to want to display the poster. By
downloading all the posters beforehand as with the rest of the Movie, there
is no waiting required and the UI will make more sense to the user.