refactor: change entities
The get_movies entity is now a simple MoviePreview with just an id and
poster.

The get_movie_details entity now has a Movies entity with all movie
details. The poster in this entity is not necessarily the same one as the
preview.

Both use cases updated with the new changes. Result looks the same from UI
perspective.